Answer "the next game" by removing the later ones from the context
Replaces the note added in aa84de8, which computed the right answer and
then asked the model to prefer it. That fixed the headline and was
overridden in the detail: Varsity volleyball still came back as September
8 with the August 20 line sitting both in the context and in the note
above it. Same result the week-range rule got from prose — its own
comment records 1-in-3 correct falling to 0-in-5.

So the later dates are taken out of the context for these questions
instead. Nothing is left to prefer because nothing else is there. That is
what the week filter's comment has always claimed it did and what it has
never actually done: `context` is assembled once and, until now, no rule
rewrote it. Only dated event lines are touched, prose is untouched, and
the trigger requires the word "next" plus a game or practice noun, so
whole-schedule questions are unaffected.

Two fixes that came out of testing it:

The weekday is spelled out in full at ingest. `(Thu)` was not enough —
given `2026-08-20 (Thu)` the model wrote "Saturday, August 20" in one
phrasing and "Tuesday, August 20 (Thu)" in another, expanding the
abbreviation and expanding it wrongly, and a prompt instruction to use the
parenthetical as written did not hold. With "Thursday" already in the line
there is nothing to expand.

The note no longer claims the lines it leaves behind are the next ones
outright. It can only sort what retrieval gave it, and the honest scope is
"among the schedules retrieved for this question".

Which is the remaining gap, stated plainly rather than left to be
rediscovered. Ask "when is the next varsity volleyball game" and the
answer is right — Thursday, August 20 at 6:00 PM vs Palmer Ridge. Ask
"when does varsity volleyball play next" and it is wrong, because that
phrasing retrieves a different chunk set which does not include the August
20 fixture, and the earliest Varsity line present is in September. The
sorting is sound; the input is not reliably complete.

Fixing that properly means not reading "next" out of the retrieved context
at all — one indexed query for the earliest upcoming event per level of
the sport asked about would be authoritative and would not depend on which
chunks happened to rank. That is the next change.

+++ src/app/api/crawl/ingest-ical/route.ts
@@ -190,9 +190,18 @@ function formatTimeOnly(dtstr: string): string {
 function formatDate(dtstr: string): string {
   const isUtc = dtstr.endsWith('Z')
   const clean = dtstr.replace(/Z$/, '')
+  /* The day is spelled out in full, not abbreviated.
+   *
+   * `(Thu)` was not enough. Given `2026-08-20 (Thu)` the model still wrote
+   * "Saturday, August 20" in one phrasing and "Tuesday, August 20 (Thu)" in
+   * another — it expands the abbreviation into prose and expands it wrongly, and
+   * a prompt instruction to use the parenthetical as written did not hold. With
+   * "Thursday" already in the line there is nothing to expand: the word the
+   * answer needs is the word that is there. Five characters per event to remove
+   * a class of error that puts a parent at a field on the wrong day. */
   const iso = (d: Date, tz: string) => {
     const ymd = d.toLocaleDateString('en-CA', { timeZone: tz })
-    const dow = d.toLocaleDateString('en-US', { weekday: 'short', timeZone: tz })
+    const dow = d.toLocaleDateString('en-US', { weekday: 'long', timeZone: tz })
     return `${ymd} (${dow})`
   }
   const m = clean.match(/(\d{4})(\d{2})(\d{2})T(\d{2})(\d{2})/)
